Responses to PR comments
