+++ .github/workflows/core.yaml
@@ -13,6 +13,7 @@ jobs:
           [
             bia-shared-datamodels,
             core,
+            bia-ro-crate,
           ]
     runs-on: ${{ matrix.os }}
     defaults:
